refactor: dedupe facade re-export boilerplate (closes #462)
Inline imports from 4 subsystem facade modules directly into their
consumers, eliminating ~142 lines of pure re-export boilerplate.

The 6 DOM facades and 4 manager barrels listed in #462 had already
been consolidated in PR #471 (into dom-facades.js and managers.js).
This change addresses the remaining 4 subsystem facades:

  - file-tree-subsystem.js     (34 lines, 3 consumers)
  - file-viewer-subsystem.js   (28 lines, 2 consumers)
  - flow-view-subsystem.js     (30 lines, 3 consumers)
  - terminal-subsystem.js      (50 lines, 2 consumers)

Each subsystem only had 2-3 consumers and added no abstraction value
beyond a thin re-export shim. Inlining preserves tree-shaking and
makes module dependencies explicit at the call site.

Verified: node build.js + npx vitest run (404/404 tests pass).
